Fix broken scroll view
Scroll view was broken on two counts.

1) The parent needs to have a flex. The scroll view is a bounded height
view rendering unbounded height children. The bounds are determined by
the space given by the parent. Flex property needs to be transfered down
the view stack in order to maintain the fixed height. Otherwise it can
lead to errors. In this case the scroll view was exceeding the bottom of
the page, meaning the content would not be entirely on-screen at the
bottom of the scroll.

2) Move padding to extra child view. Padding on scroll view caused
issues. Because reasons.